Add controllers, entities, dtos, enums, middleware, and other necessary files
A large number of scripts have been added in this commit. Important additions include controllers for companies, entities for companies and agencies, and data transfer objects for companies. Simple npm dependencies and scripts for the working of TypeORM have also been added. Other supporting implementations for the functionality make up the rest of the changes.